Separate extensions as their own files
Give each extension its own test file. Primarily this is done to make
it easier to add some fixtures and extend the geo tests.
